+++ qsprpred/extra/gpu/models/chemprop.py
@@ -14,8 +14,9 @@ from tqdm import trange
 
 from ....data.data import QSPRDataset
 from ....models.early_stopping import EarlyStoppingMode, early_stopping
-from ....models.interfaces import QSPRModel
+from ....models.interfaces import QSPRModel, FitMonitor
 from ....models.tasks import ModelTasks
+from ....models.monitors import BaseMonitor
 
 
 class ChempropMoleculeModel(chemprop.models.MoleculeModel):
@@ -167,6 +168,7 @@ class ChempropModel(QSPRModel):
         y: pd.DataFrame | np.ndarray | QSPRDataset,
         estimator: Any = None,
         mode: EarlyStoppingMode = EarlyStoppingMode.NOT_RECORDING,
+        monitor: FitMonitor | None = None,
         keep_logs: bool = False,
     ) -> Any | tuple[ChempropMoleculeModel, int | None]:
         """Fit the model to the given data matrix or `QSPRDataset`.
@@ -181,15 +183,18 @@ class ChempropModel(QSPRModel):
             X (pd.DataFrame, np.ndarray, QSPRDataset): data matrix to fit
             y (pd.DataFrame, np.ndarray, QSPRDataset): target matrix to fit
             estimator (Any): estimator instance to use for fitting
-            early_stopping (bool): if True, early stopping is used,
-                                   only applies to models that support early stopping.
+            mode (EarlyStoppingMode): mode to use for early stopping
+            monitor (FitMonitor): monitor to use for fitting, if None, a BaseMonitor
+                is used
 
         Returns:
             Any: fitted estimator instance
             int: in case of early stopping, the number of iterations
                 after which the model stopped training
         """
+        monitor = BaseMonitor() if monitor is None else monitor
         estimator = self.estimator if estimator is None else estimator
+        monitor.onFitStart(self)
 
         # convert data to chemprop MoleculeDataset
         data = self.convertToMoleculeDataset(X, y)
@@ -309,6 +314,7 @@ class ChempropModel(QSPRModel):
             self.earlyStopping.getEpochs() if not self.earlyStopping else args.epochs
         )
         for epoch in trange(n_epochs):
+            monitor.onEpochStart(epoch)
             self.chempropLogger.debug(f"Epoch {epoch}")
             n_iter = chemprop.train.train(
                 model=estimator,
@@ -360,6 +366,7 @@ class ChempropModel(QSPRModel):
                 mean_val_score = chemprop.utils.multitask_mean(
                     val_scores[args.metric], metric=args.metric
                 )
+                monitor.onEpochEnd(epoch, mean_val_score)
                 if (
                     args.minimize_score and mean_val_score < best_score or
                     not args.minimize_score and mean_val_score > best_score
@@ -378,7 +385,9 @@ class ChempropModel(QSPRModel):
             shutil.rmtree(save_dir)
 
         if self.earlyStopping:
+            monitor.onFitEnd(best_estimator, best_epoch)
             return best_estimator, best_epoch
+        monitor.onFitEnd(estimator)
         return estimator, None
 
     def predict(