+++ qsprpred/models/early_stopping.py
@@ -173,6 +173,7 @@ def early_stopping(func: Callable) -> Callable:
         y: pd.DataFrame | np.ndarray | QSPRDataset,
         estimator: Any | None = None,
         mode: EarlyStoppingMode | None = None,
+        monitor: "FitMonitor" = None,
         **kwargs,
     ) -> Any:
         """Wrapper for fit method of models that support early stopping.
@@ -182,6 +183,8 @@ def early_stopping(func: Callable) -> Callable:
             y (pd.DataFrame, np.ndarray, QSPRDataset): target matrix to fit
             estimator (Any): estimator instance to use for fitting
             mode (EarlyStoppingMode): early stopping mode
+            monitor (FitMonitor): monitor to use for fitting, if None, a BaseMonitor
+                is used
             kwargs (dict): additional keyword arguments for the estimator's fit method
 
         Returns:
@@ -192,7 +195,7 @@ def early_stopping(func: Callable) -> Callable:
             " early stopping."
         )
         self.earlyStopping.mode = mode if mode is not None else self.earlyStopping.mode
-        estimator, best_epoch = func(self, X, y, estimator, mode, **kwargs)
+        estimator, best_epoch = func(self, X, y, estimator, mode, monitor, **kwargs)
         if self.earlyStopping.mode == EarlyStoppingMode.RECORDING:
             self.earlyStopping.recordEpochs(best_epoch + 1)  # +1 for 0-indexing
         return estimator